src/main/java/ru/education/controllers/ProductController.java
package ru.education.controllers;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.security.access.prepost.PreAuthorize;
import org.springframework.web.bind.annotation.*;
import ru.education.entity.Product;
import ru.education.service.ProductService;

import java.util.List;

@RestController
@RequestMapping("api/v1/product")
public class ProductController {

    private final ProductService productService;

    public ProductController(ProductService productService) {
        this.productService = productService;
    }

    @GetMapping
    @PreAuthorize("hasPermission('product', 'read')")
    public List<Product> findAll(){
        return productService.findAll();
    }

    @GetMapping("/{id}")
    @PreAuthorize("hasPermission('product', 'readById')")
    public Product findById(@PathVariable String id){
        return productService.findById(id);
    }

    @PostMapping
    @ResponseStatus(HttpStatus.CREATED)
    @PreAuthorize("hasPermission('product', 'create')")
    public Product create(@RequestBody Product product){
        return productService.create(product);
    }

    @PutMapping
    @ResponseStatus(HttpStatus.OK)
    @PreAuthorize("hasPermission('product', 'update')")
    public Product update(@RequestBody Product product){
        return productService.update(product);
    }

    @DeleteMapping("/{id}")
    @ResponseStatus(HttpStatus.NO_CONTENT)
    @PreAuthorize("hasPermission('product', 'delete')")
    public void delete(@PathVariable String id){
        productService.delete(id);
    }
}
